# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Optimization Dashboard API
  description: |
    Comprehensive API for portfolio optimization using modern portfolio theory and advanced financial algorithms.
    
    ## Features
    - **Portfolio Optimization**: Markowitz, Black-Litterman, Risk Parity algorithms
    - **Risk Analytics**: VaR, CVaR, Sharpe ratio, maximum drawdown calculations
    - **Monte Carlo Simulations**: Up to 100,000 scenario analysis
    - **Backtesting**: Historical performance validation
    - **Real-time Data**: Market data integration with caching
    - **Security**: JWT authentication, rate limiting, input validation
    
    ## Authentication
    This API uses JWT (JSON Web Tokens) for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limits
    - Authentication endpoints: 5 requests per 5 minutes
    - General API endpoints: 100 requests per minute  
    - Optimization endpoints: 10 requests per minute
    - File upload endpoints: 5 requests per minute
    
    ## Error Handling
    The API uses standard HTTP status codes and returns detailed error messages in JSON format.
    
  version: "1.0.0"
  contact:
    name: Portfolio Optimization Team
    email: api-support@portfolio-dashboard.com
    url: https://docs.portfolio-dashboard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.portfolio-dashboard.com/v1
...
security:
  - BearerAuth: []
paths:
...
  /optimization/optimize:
    post:
      tags:
        - Optimization
      summary: Optimize portfolio
      description: |
        Run portfolio optimization using specified algorithm and constraints.
        
        Supported optimization methods:
        - **max_sharpe**: Maximize Sharpe ratio (risk-adjusted returns)
        - **min_volatility**: Minimize portfolio volatility
        - **max_return**: Maximize expected returns (with risk constraints)
        - **risk_parity**: Equal risk contribution from all assets
        - **target_return**: Minimize risk for target return level
        - **black_litterman**: Bayesian optimization with market views
        
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OptimizationRequest'
            examples:
              max_sharpe:
                summary: Maximum Sharpe Ratio
                value:
                  symbols: ["AAPL", "GOOGL", "MSFT", "TSLA"]
                  method: "max_sharpe"
                  risk_free_rate: 0.02
                  constraints:
                    max_weight: 0.4
                    min_weight: 0.05
              risk_parity:
                summary: Risk Parity Portfolio
                value:
                  symbols: ["SPY", "TLT", "GLD", "VTI"]
                  method: "risk_parity"
                  constraints:
                    max_weight: 0.5
      responses:
        '200':
          description: Optimization completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptimizationResult'
        '400':
          description: Invalid optimization parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Optimization failed (infeasible constraints)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /optimization/efficient-frontier:
    post:
      tags:
        - Optimization
      summary: Generate efficient frontier
      description: Generate risk-return efficient frontier points for visualization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                symbols:
                  type: array
                  items:
                    type: string
                  minItems: 2
                  maxItems: 50
                  example: ["AAPL", "GOOGL", "MSFT", "AMZN"]
                num_points:
                  type: integer
                  minimum: 10
                  maximum: 200
                  default: 50
                  description: Number of frontier points to generate
                start_date:
                  type: string
                  format: date
                  description: Historical data start date
                  example: "2020-01-01"
                end_date:
                  type: string
                  format: date
                  description: Historical data end date
                  example: "2023-12-31"
                constraints:
                  $ref: '#/components/schemas/OptimizationConstraints'
              required:
                - symbols
      responses:
        '200':
          description: Efficient frontier generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EfficientFrontierResponse'
...
  /optimization/backtest:
    post:
      tags:
        - Optimization
      summary: Backtest portfolio strategy
      description: |
        Perform historical backtesting of portfolio optimization strategy.
        Tests how the optimization would have performed with historical data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BacktestRequest'
      responses:
        '200':
          description: Backtest completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestResult'

  # Market Data
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for API authentication
  schemas:
...
    OptimizationRequest:
      type: object
      properties:
        symbols:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 50
        method:
          type: string
          enum: [max_sharpe, min_volatility, max_return, risk_parity, target_return, black_litterman]
        risk_free_rate:
          type: number
          format: float
          default: 0.02
          minimum: 0
          maximum: 0.1
        target_return:
          type: number
          format: float
          description: Required for target_return method
        start_date:
          type: string
          format: date
          description: Historical data start date
        end_date:
          type: string
          format: date
          description: Historical data end date
        constraints:
          $ref: '#/components/schemas/OptimizationConstraints'
        market_views:
          type: array
          items:
            $ref: '#/components/schemas/MarketView'
          description: Required for black_litterman method
      required:
        - symbols
        - method
    OptimizationConstraints:
      type: object
      properties:
        max_weight:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Maximum weight for any asset
        min_weight:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Minimum weight for any asset
        max_volatility:
          type: number
          format: float
          minimum: 0
          description: Maximum portfolio volatility
        target_volatility:
          type: number
          format: float
          minimum: 0
          description: Target portfolio volatility
        sector_limits:
          type: object
          additionalProperties:
            type: number
            format: float
            minimum: 0
            maximum: 1
          description: Maximum allocation per sector
    MarketView:
      type: object
      properties:
        asset:
          type: string
          description: Asset symbol
        expected_return:
          type: number
          format: float
          description: Expected return (annual)
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence level (0-1)
      required:
        - asset
        - expected_return
        - confidence
    OptimizationResult:
      type: object
      properties:
        weights:
          type: object
          additionalProperties:
            type: number
            format: float
          description: Optimal asset weights
        expected_return:
          type: number
          format: float
          description: Expected annual return
        volatility:
          type: number
          format: float
          description: Portfolio volatility (annual)
        sharpe_ratio:
          type: number
          format: float
          description: Risk-adjusted return metric
        max_drawdown:
          type: number
          format: float
          description: Maximum historical drawdown
        var_95:
          type: number
          format: float
          description: Value at Risk (95% confidence)
        cvar_95:
          type: number
          format: float
          description: Conditional Value at Risk (95% confidence)
        computation_time:
          type: number
          format: float
          description: Optimization time in seconds
        converged:
          type: boolean
          description: Whether optimization converged
        objective_value:
          type: number
          format: float
          description: Objective function value

    # Additional schemas continue...
    EfficientFrontierResponse:
      type: object
      properties:
        frontier_points:
          type: array
          items:
            type: object
            properties:
              expected_return:
                type: number
                format: float
              volatility:
                type: number
                format: float
              weights:
                type: object
                additionalProperties:
                  type: number
                  format: float
        optimal_portfolio:
          $ref: '#/components/schemas/OptimizationResult'
        computation_time:
          type: number
          format: float
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time

    # More schemas for completeness...
...